v1.51: Arch Linux compatibility + cross-platform path fixes
setup.sh:
  - Fix unterminated echo string (line 55) that crashed setup with a
    bash parse error before any venv was created.
  - Switch to uv-managed venvs (uv venv --seed, uv pip install) so the
    'No module named pip' error after a fresh uv venv is gone.
  - Detect Arch (pacman) and Debian/Ubuntu (apt) and install the right
    system packages: portaudio, libsndfile, ffmpeg, libnotify, pipewire.
  - Auto-download Qwen3-0.6B into assistant/model/cozy-llm-v1/ so
    'cozy' works on first launch without a separate training run.
  - Derive the shell-alias source path from $ROOT instead of
    hardcoding /home/saptodeepsarkar (works for any user).

Path fixes (12 files):
  - assistant/tui_textual.py:791   hardcoded WW_PATH -> relative
  - cozy-vision/harness/cozy_swap.py  /home/saptodeepsarkar -> COZY_ROOT
  - team/scripts/{retrain_clean,retrain_cpu,train_lr_fast}.py
                                   hardcoded WW -> Path(__file__).parents[2]
  - team/scripts/extract_segment.py  hardcoded open() -> Path-relative
  - assistant/model/cozy-llm-v1-dpo/README.md
                                   base_model paths -> relative
  - AGENTS.md, cozy.shell, stt-finetune/.../manifest.jsonl
                                   rebased

Code fixes:
  - stt-finetune/scripts/train_me.py:111  SyntaxError: unmatched ')'
  - 13 stt-finetune scripts: shebang python -> python3 (Arch has no
    /usr/bin/python)
  - assistant/rlm_harness/{harness_fast,tui,plugins/{llm,stt,tts,wake}}.py
                                   absolute 'from rlm_harness.X' -> relative
  - 7 from_pretrained(...) calls: dtype= -> torch_dtype= (compat with
    transformers 4.x and 5.x)
  - wakeword/src/livekit/wakeword/data/  restub package; lazy CLI
    imports no longer raise ModuleNotFoundError on a clean checkout.
  - cozy-vision/setup.sh  add Arch (pacman) branch.

executor.py: Wayland/PipeWire-first fallbacks:
  - volume:  wpctl -> pactl -> amixer
  - shot:    grim -> gnome-screenshot -> scrot -> import
  - lock:    loginctl -> swaylock -> hyprlock -> waylock -> ...
  - APP_ALIASES: add thunar, dolphin, konsole, alacritty, kate,
    qalculate-gtk, claws-mail, etc.

cozy: STT models no longer block launch (CozySTT is lazy); missing
files now print a 'Note (optional, lazy-loaded)' hint and continue.

+++ assistant/rlm_harness/harness.py
@@ -132,13 +132,13 @@ class ModelBackend(Backend):
             # base + LoRA adapter
             from peft import PeftModel
             self._model = AutoModelForCausalLM.from_pretrained(
-                str(base), dtype=torch_dtype,
+                str(base), torch_dtype=torch_dtype,
                 attn_implementation="sdpa")
             self._model = PeftModel.from_pretrained(
                 self._model, str(Path(adapter_dir)))
         else:
             self._model = AutoModelForCausalLM.from_pretrained(
-                str(base), dtype=torch_dtype,
+                str(base), torch_dtype=torch_dtype,
                 attn_implementation="sdpa")
 
         self._model = self._model.to(device)
